Minor cleanup: reflect that UserData is not gzipped.
Contrary to the online docs, and for some reason, Ubuntu CloudInit
won't recognize the UserData as an executable to run if it's gzipped.

This should be no problem since the payload (the bootstrap tree) is
bz2-compressed and we have plenty of room left before we hit the
UserData size limit.