First shot at replacing ugly hack for initializing hourly cron job.
I'm getting the minute from a random number seeded by a hash of the
server hostname, so hopefully instances will not place too much
stress upon any common resource they might need to update
themselves.

I was initializing this with a Python script, feeding the result
into the masterless/init.sls on bootstrap.  This was ugly and
caused the masterless module to report errors ever after, since
this hack was not done for every update.  Now I do it the proper
way, using a Python script as the definition for that part of
the module.

This is as of yet untested.  I'm checking it in so I can push it
to a running instance for debugging.