Add bootstrap module and public proxy port setup file.
The bootstrap module will be disabled in all branches but master,
which is the one where instances are supposed to be launched.

At this moment it only copies a file containing the public proxy
port of this instance to /etc/lantern.  User and permissions for
this directory and file will be set in the lantern module.

I do this in a bootstrap module because the public proxy port is
instance specific and decided at launch time.  It would be
cumbersome and pointless to keep managing this file in salt for
all launched nodes.