Pin boto version.
It's good for stability in general to have well-known versions of stuff.
In this case, we want to have a known shared version between cloudmaster
and newly launched instances because we will want to share pickled
boto Messages between them.  This is so a launched instance can delete
a SQS that the cloudmaster has fetched.