Avoid any characters but alphanumeric in instance names.
I got a problem by which a period in instance names was confusing
the pillar system.

Our use of base64 here is a remnant of when we were generating the
instance name as a hash of the user email address.  We have a
different approach to mapping emails to instance names now, so let's
use the safer, more straightforward approach of just generating a
random alphanumeric string.