Change AWS credentials, make them configured by pillars.
The old AWS credentials were leaked a few minutes ago so I had to
delete them and create new ones.  They used to be configured in
Salt instances as grains, which are only ever supposed to be
initialized on machine startup.  I've changed them to pillars,
which is the preferred mechanism for reconfigurable data.