Remove tracking of instance name by e-mail.
Currently this was only needed for some rally-related use cases.
I'm removing all the rally-specific code for simplicity.  When
the dust settles on the fallback-balancing branch and we come
back to rally, I'll add this back and adapt it to the new scheme
where users have possibly more than one fallback proxy running
as them.