Delete .littleproxy too.
I initially did this for the same reason I deleted the .lantern
folder -- that is, to avoid any state buildup between invocations
of the job that checks that we can proxy through all fallbacks.

Later I found out that the original patch had regressed, but I
keep this commit since it can't hurt anyway.