fix path to install-lantern-from-git.bash
This was broken when moving this from fallback_proxy to a common
.sls that could be shared with the cloudmaster.  It worked for me
locally because somehow I still had the file in the old location.

Also, remove the renaming of check_lantern.py.  It is no longer the
case that it will kill all java processes, so it shouldn't interfere
with the building of Lantern.