step back on changes that would break our setup
No released Salt version supports Digital Ocean's 2.0 API, although github
HEAD apparently does.

Ubuntu 14.04 will fail in cloudmasters due to some iptables problem.  12.04
has been working fine for us so there's no urgency to fix this.  For some
reason, fallbacks themselves aren't affected by this problem, so we build them
with version 14.04.  This is not much of a problem since there are other
ways in which the cloudmaster's setup is singular.

This was tested by launching a new cloudmaster and a fallback proxy.