Make sure we force a Salt reinstall if we don't have a salt-minion executable
The combination of the salt bootstrap script and our removal of the salt-minion
apt-get package may get us into a state where we have a valid Salt library
(e.g., salt-call) but no salt-minion executable.  Make sure we reinstall from
scratch whenever we get a salt-minion.

Also remove the salt-nodeps state, since it doesn't hurt to have the most
up-to-date versions of the dependencies even when not strictly necessary.
I don't remember whether having too new a version of a dependency has caused
problems in the past, but that is not the case currently.