make upgrading salt-minion more robust
Ensure that we have removed the old salt-minion when we check whether we need
to reinstall Salt to get the new one.

Make this all happen in a process outside of the old salt-minion.

Rebooting after initial setup is desirable in general since sometimes an
apt-get upgrade will put the system in a state where a reboot is required, but
that won't happen until triggered by unattended upgrades.  More apropos to this
commit, it makes sure the old salt-minion service is killed.