stop trying to manage salt version through a salt configuration
The main reason for doing this is that at some point I knew of no other way to
keep salt versions stable enough and in sync between the master and minions.
At this moment, this can be done by just installing from a specific git version
(which salt-cloud and salt-bootstrap allow us to do nowadays) and only ever
upgrading manually via ssh.