first (failed) shot at solving #140
This won't work because offloading the servers requires bringing
the clientip->srv table into RAM, which will swamp a regular proxy.

I'm saving this for future reference, and move on to a solution
where the proxy requests that the cloudmaster does the job, much
like how retiring is being done now.